# server/openapi.yaml
openapi: 3.0.3
info:
  title: LEGION
  version: 0.1.0
  description: The engine. One Django server, every limb a client of it.
paths:
  /api/auth/login:
    post:
      operationId: api_auth_login_create
      description: |-
        `POST /api/auth/login`. Rate-limited 5/min per IP so a leaked email
        cannot be brute-forced into a working password - the household is two
        people, not a userbase, so five attempts a minute is generous, not
        tight.
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
          description: 'A new device token. `token` is the raw key and is shown ONCE
            - the server stores only its hash and cannot hand it back again. Send
            it as `Authorization: Token <key>` on every later call.'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detail'
          description: Wrong email or password, or the account is inactive. No token
            was issued.
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detail'
          description: 'Rate limited: 5 attempts a minute per IP. `detail` says when
            to try again.'
  /api/auth/logout:
    post:
      operationId: api_auth_logout_create
      description: |-
        `POST /api/auth/logout`. Revokes only the token making the request -
        every other device this user owns keeps working.

        Send it with an explicit `Content-Length: 0` even though there is no
        body: measured against the deployed Cloud Run service on 2026-09-07, a
        POST carrying neither a body nor that header is answered 411 Length
        Required by the frontend and never reaches Django, so the token is NOT
        revoked and the caller has been told nothing about why. Most HTTP
        clients set the header themselves for an empty POST; `curl` without
        `-d` does not.
      tags:
      - auth
      security:
      - deviceToken: []
      responses:
        '204':
          description: This token is revoked. Every other device keeps working.
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detail'
          description: The request carried no device token, so there was nothing to
            revoke. Nothing was changed.
  /api/auth/me:
    get:
      operationId: api_auth_me_retrieve
      description: |-
        `GET /api/auth/me`. This is the phone's membership check - if this
        call succeeds, the calling token is live and its user is a household
        member; if it 401s or 403s, the phone knows to ask for a new token.
      tags:
      - auth
      security:
      - deviceToken: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
          description: The token is live and its user is a household member. `device_name`
            is the name the token was issued under at login.
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detail'
          description: No token, an unknown token, or a revoked one. Ask for a new
            one.
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detail'
          description: The token is live but its user is not a household member. A
            `User` row alone is not enough - see `manage.py add_household_member`.
...
components:
  schemas:
...
    Detail:
      type: object
      description: |-
        `{"detail": "..."}` - the shape every hand-written refusal in this API
        uses, and the shape DRF itself uses for 401/403/404/405. The strings are
        written to be shown to a person: they say what did NOT happen (CLAUDE.md
        section 7's outcome-verb rule), so a client should surface `detail`
        rather than inventing its own wording.
      properties:
        detail:
          type: string
      required:
      - detail
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          writeOnly: true
        device_name:
          type: string
          maxLength: 255
      required:
      - device_name
      - email
      - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user_id:
          type: string
          format: uuid
      required:
      - token
      - user_id
    MeResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        device_name:
          type: string
      required:
      - device_name
      - email
      - user_id
...
  securitySchemes:
    deviceToken:
      type: apiKey
      in: header
      name: Authorization
      description: One token per device (phone, browser, future robot), revocable
        alone. Send the value as `Token <key>`, including the word Token and the space.
        Get a key from POST /api/auth/login; POST /api/auth/logout revokes only the
        key that made the request.